+++ src/TestSuiteNotCreatedException.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace MyTester;
+
+/**
+ * Exception thrown when a test suite could not be created
+ *
+ * @author Jakub Konečný
+ */
+class TestSuiteNotCreatedException extends \RuntimeException
+{
+}