+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 8.1.0-dev
 - allowed using different methods for running code before/after test case/method
 - deprecated Job::$onAfterExecute in favor of new event TestJobFinished
+- allowed returning null from ITestSuiteFactory::create()
+- it is possible to use multiple test suite factories with ChainTestSuiteFactory
 
 Version 8.0.0
 - dropped support for Composer 2.0 and 2.1